listo para revisar

+++ src/com/demo/model/entity/Usuario.java
@@ -1,17 +1,13 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package com.demo.model.entity;
 
 /**
  *
- * @author whiston
+ * @author Fekilo
  */
 public class Usuario {
     
-    private long id;
+    private int id;
     private String email;
     private String clave;
     private String apellidos;
@@ -21,7 +17,7 @@ public class Usuario {
     public Usuario() {
     }
 
-    public Usuario(long id, String email, String clave, String apellidos, String nombres, int tipo) {
+    public Usuario(int id, String email, String clave, String apellidos, String nombres, int tipo) {
         this.id = id;
         this.email = email;
         this.clave = clave;
@@ -35,11 +31,11 @@ public class Usuario {
         this.clave = clave;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
